Move `check_balance_changes` to Bridge pool VP module
